move adLayout back to adapter.
