+++ ThirdPartyAdapters/vungle/vungle/src/main/java/com/vungle/mediation/VungleInterstitialAdapter.java
@@ -19,6 +19,7 @@ import android.os.Bundle;
 import android.text.TextUtils;
 import android.util.Log;
 import android.view.View;
+import android.widget.RelativeLayout;
 import androidx.annotation.Keep;
 import com.google.ads.mediation.vungle.VungleInitializer;
 import com.google.android.gms.ads.AdRequest;
@@ -44,6 +45,10 @@ public class VungleInterstitialAdapter
   private AdConfig mAdConfig;
   private String mPlacementForPlay;
 
+  /**
+   * Ad container for Vungle's banner ad.
+   */
+  private volatile RelativeLayout adLayout;
   private VungleBannerAdapter vungleBannerAdapter;
 
   @Override
@@ -204,18 +209,46 @@ public class VungleInterstitialAdapter
   public void requestBannerAd(Context context, MediationBannerListener mediationBannerListener,
       Bundle serverParameters, AdSize adSize, MediationAdRequest mediationAdRequest,
       Bundle mediationExtras) {
+    // Create the adLayout wrapper with the requested ad size, as Vungle's ad uses MATCH_PARENT for
+    // its dimensions.
+    adLayout =
+        new RelativeLayout(context) {
+          @Override
+          protected void onAttachedToWindow() {
+            super.onAttachedToWindow();
+            if (vungleBannerAdapter != null) {
+              vungleBannerAdapter.attach();
+            }
+          }
+
+          @Override
+          protected void onDetachedFromWindow() {
+            super.onDetachedFromWindow();
+            if (vungleBannerAdapter != null) {
+              vungleBannerAdapter.detach();
+            }
+          }
+        };
+    int adLayoutHeight = adSize.getHeightInPixels(context);
+    // If the height is 0 (e.g. for inline adaptive banner requests), use the closest supported size
+    // as the height of the adLayout wrapper.
+    if (adLayoutHeight <= 0) {
+      float density = context.getResources().getDisplayMetrics().density;
+      adLayoutHeight = Math.round(mAdConfig.getAdSize().getHeight() * density);
+    }
+    RelativeLayout.LayoutParams adViewLayoutParams =
+        new RelativeLayout.LayoutParams(adSize.getWidthInPixels(context), adLayoutHeight);
+    adLayout.setLayoutParams(adViewLayoutParams);
+
     vungleBannerAdapter = new VungleBannerAdapter(context, VungleInterstitialAdapter.this,
-        mediationBannerListener);
+        mediationBannerListener, adLayout);
     vungleBannerAdapter
         .requestBannerAd(adSize, mediationAdRequest, serverParameters, mediationExtras);
   }
 
   @Override
   public View getBannerView() {
-    if (vungleBannerAdapter != null) {
-      return vungleBannerAdapter.getBannerView();
-    }
-    return null;
+    return adLayout;
   }
 
 }
